In this commit , the type conversion feature is added into the code , so that compiler automatically convert one datatype to another. Right now only int to float and float to int type conversion is included.

+++ test.c
@@ -8,7 +8,7 @@ float calculate_area(float radius){
 
 int main(){
     int radius = 5;
-    int area = radius*radius;
+    float area = radius*radius;
 
     int a = area/4;
